+++ packages/insomnia/src/ui/insomnia-components/index.ts
@@ -4,7 +4,6 @@ export * from './list-group';
 
 export { CardContainer } from './card-container';
 export { Card, type  CardProps } from './card';
-export { Header, type  HeaderProps } from './header';
 export { MultiSwitch } from './multi-switch';
 export { RadioButtonGroup, type RadioButtonGroupProps } from './radio-button-group';
 export { SvgIcon, type SvgIconProps, type IconEnum, type ThemeEnum, type IconId } from './svg-icon';
